requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pymongo
pydantic
